docs:Adding documentation
to explain how to install and configure Wasp.
The docs are based on the v1.0 documentation
and have been adjusted to reflect the Go
rewrite, as well as the new eviction logic
and limited swap features added to the wasp-agent.

Also added section for upgrade procedure from v1.0 to
the current one.

Signed-off-by: bmordeha <bmordeha@redhat.com>